+++ src/OverflowScroll/data/index.js
@@ -1,5 +1,6 @@
 export * from './constants';
 
+export { default as useOverflowScroll } from './useOverflowScroll';
 export { default as useOverflowScrollActions } from './useOverflowScrollActions';
 export { default as useOverflowScrollElementAttributes } from './useOverflowScrollElementAttributes';
 export { default as useOverflowScrollEventListeners } from './useOverflowScrollEventListeners';
