+++ src/OverflowScroll/index.js
@@ -1,4 +1,6 @@
 export { default as OverflowScroll } from './OverflowScroll';
 export { default as OverflowScrollContext } from './OverflowScrollContext';
-export { default as useOverflowScroll } from './useOverflowScroll';
-export { default as useOverflowScrollItems } from './data/useOverflowScrollItems';
+export {
+  useOverflowScroll,
+  useOverflowScrollItems,
+} from './data';
